Robot Perception (ROB-GY 6203)/Midterm-project/requirements.txt
joblib==1.4.2
matplotlib==3.5.1
natsort==8.4.0
numpy==1.21.5
opencv_contrib_python==4.10.0.84
pandas==2.2.3
Pillow==9.0.1
Pillow==11.0.0
pygame==2.6.1
scikit_learn==1.5.2
tensorflow==2.18.0
tqdm==4.66.4
